requirements.txt
pytest~=7.2.2
selene~=2.0.0b16
python-dotenv~=0.21.1
allure-pytest~=2.12.0
